+++ docs/src/markdown/about/changelog.md
@@ -3,6 +3,11 @@ icon: lucide/scroll-text
 ---
 # Changelog
 
+## 8.8
+
+-   **NEW**: Add new spaces: `helmlab`, `helmgen`, and `helmgenlch`.
+-   **NEW**: Add new distancing algorithm: `helmlab`.
+
 ## 8.7
 
 -   **NEW**: Drop support for Python 3.9.